+++ src/main/java/com/yohanii/lostandfound/dto/profile/ProfileEditRequestDto.java
@@ -1,0 +1,15 @@
+package com.yohanii.lostandfound.dto.profile;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class ProfileEditRequestDto {
+
+    private String name;
+    private String nickName;
+
+}